update changes project Karate PetStore
