+++ PetStoreKarate/src/test/java/karate/AllfeatureTest.java
@@ -1,6 +1,9 @@
 package karate;
+
+import com.intuit.karate.KarateOptions;
 import com.intuit.karate.Results;
 import com.intuit.karate.Runner;
+import com.intuit.karate.junit5.Karate;
 import net.masterthought.cucumber.Configuration;
 import net.masterthought.cucumber.ReportBuilder;
 import org.apache.commons.io.FileUtils;
@@ -11,7 +14,7 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-
+@KarateOptions(tags = {"@LoginUser"})
 public class AllfeatureTest {
     public static void generateReport(String karateOutputPath){
         Collection<File> jsonFiles = FileUtils.listFiles(new File(karateOutputPath), new String[]{"json"}, true);
