+++ PetStoreKarate/src/test/resources/Karate/request/Pet/1AddnewPet/requestBodyAdd.json
