+++ PetStoreKarate/src/test/resources/Karate/request/Pet/3FindPet/responseFind.json
