+++ PetStoreKarate/src/test/resources/Karate/request/Pet/4DeletePet/4DeletePet.feature
