+++ PetStoreKarate/src/test/resources/Karate/request/User/5CreateUser/requestBodyLogin.json
@@ -1,0 +1,10 @@
+{
+  "id": 1,
+  "username": "Nad",
+  "firstName": "First",
+  "lastName": "Last",
+  "email": "nadia123mickey@hotmail.com",
+  "password": "12df3df3456",
+  "phone": "5537858126",
+  "userStatus": 1
+}
