+++ PetStoreKarate/src/test/resources/Karate/request/User/5CreateUser/responseLogin.json
@@ -1,0 +1,5 @@
+{
+  "code": "#number",
+  "type": "#string",
+  "message": "#string"
+}
